services/python/pyproject.toml
[project]
name = "python-service"
version = "0.1.0"
description = "Microserviço Python para processamento de dados de séries temporais."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.8",
    "avrotize>=3.0.2",
    "pika>=1.3.2",
    "protobuf-to-pydantic>=0.3.3.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "betterproto[compiler]>=2.0.0b6",
    "grpcio-tools>=1.62.0",
    "prophet>=1.1.5",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["scripts", "models"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "datamodel-code-generator>=0.53.0",
    "fastapi>=0.128.0",
    "hatchling>=1.28.0",
    "uvicorn>=0.40.0",
    "watchdog>=6.0.0",
    "ruff>=0.9.0",
]

[project.scripts]
gen = "scripts.codegen:main"
dev = "scripts.dev:main"

[tool.ruff]
line-length = 120
target-version = "py311"
indent-width = 2

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (imports)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ANN",    # flake8-annotations (type hints)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "ANN101", # missing type annotation for self (deprecated)
    "ANN102", # missing type annotation for cls (deprecated)
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = false
suppress-none-returning = true

[tool.ruff.lint.isort]
known-first-party = ["models", "scripts", "services"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
